Moved middleware
