+++ example/main.rs
@@ -1,14 +1,10 @@
-use futures::{Future, Stream};
 use serde_derive::*;
-use serde_json::Value;
-use url::form_urlencoded;
 
 use obsidian::{
-    context::Context,
     header,
-    middleware::Middleware,
+    middleware::{BodyParser, Logger, UrlEncodedParser},
     router::{RequestData, ResponseBuilder},
-    App, Body, Request, Response, StatusCode,
+    App, StatusCode,
 };
 
 // Testing example
@@ -16,108 +12,6 @@ use obsidian::{
 struct Point {
     x: i32,
     y: i32,
-}
-
-pub struct BodyParser {}
-pub struct UrlEncodedParser {}
-pub struct Logger {}
-
-impl BodyParser {
-    pub fn new() -> Self {
-        BodyParser {}
-    }
-}
-
-impl Middleware for BodyParser {
-    fn handle<'a>(
-        &'a self,
-        context: Context<'a>,
-    ) -> Box<Future<Item = Response<Body>, Error = hyper::Error> + Send> {
-        let mut context = context;
-        let (parts, body) = context.request.into_parts();
-
-        let b = match body.concat2().wait() {
-            Ok(chunk) => chunk,
-            Err(e) => {
-                println!("{}", e);
-                hyper::Chunk::default()
-            }
-        };
-
-        let json_result: serde_json::Result<Value> = serde_json::from_slice(&b);
-
-        match json_result {
-            Ok(json_body) => context.route_data.add_json(json_body),
-            Err(e) => println!("{}", e),
-        }
-
-        let req = Request::from_parts(parts, Body::from(b));
-
-        context.request = req;
-        context.next()
-    }
-}
-
-impl UrlEncodedParser {
-    pub fn new() -> Self {
-        UrlEncodedParser {}
-    }
-}
-
-impl Middleware for UrlEncodedParser {
-    fn handle<'a>(
-        &'a self,
-        context: Context<'a>,
-    ) -> Box<Future<Item = Response<Body>, Error = hyper::Error> + Send> {
-        let mut context = context;
-        let (parts, body) = context.request.into_parts();
-
-        let b = match body.concat2().wait() {
-            Ok(chunk) => chunk,
-            Err(e) => {
-                println!("{}", e);
-                hyper::Chunk::default()
-            }
-        };
-
-        let params_iter = form_urlencoded::parse(b.as_ref()).into_owned();
-
-        for (key, value) in params_iter {
-            context.route_data.add_param(key, value);
-        }
-
-        let req = Request::from_parts(parts, Body::from(b));
-
-        context.request = req;
-        context.next()
-    }
-}
-
-impl Logger {
-    pub fn new() -> Self {
-        Logger {}
-    }
-}
-
-impl Middleware for Logger {
-    fn handle<'a>(
-        &'a self,
-        context: Context<'a>,
-    ) -> Box<Future<Item = Response<Body>, Error = hyper::Error> + Send> {
-        println!(
-            "{} {} \n{}",
-            context.request.method(),
-            context.request.uri(),
-            context
-                .request
-                .headers()
-                .get("host")
-                .unwrap()
-                .to_str()
-                .unwrap()
-        );
-        context.next()
-    }
 }
 
 fn main() {
